app/requirements.txt
httpx
selenium
pydantic
pypdf
webdriver-manager
chromedriver_binary
selenium-wire
google-api-python-client
google-auth

azure-servicebus
azure-identity
azure-storage-blob
aiohttp
python-dotenv
blinker==1.7.0
